### 2018-03-25 #646 @whedon
[PRE REVIEW]: Spiral: splitters for identifiers in source code files
**Submitting author:** @mhucka (<a href="http://orcid.org/0000-0001-9105-5960">Michael Hucka</a>)
**Repository:** <a href="https://github.com/casics/spiral" target ="_blank">https://github.com/casics/spiral</a>
**Version:** 1.0.0
