+++ tests/README.md
@@ -1,6 +1,15 @@
 Tests
 =====
 
+Tests for `py.test`
+-------------------
+
+The files whose names begin with `test_` are intended for use with the [py.test](https://pytest.org) testing framework.  You can run them by executing the following command in a terminal shell in this directory:
+
+```csh
+python3 -m py.test
+```
+
 `compare`
 ---------
 
@@ -14,12 +23,3 @@ The small program named `compare` in this directory is a utility I used during t
 The `-l` option is needed with the INTT data set because in that set, all the identifiers have been lower-cased in the expected results list.  The `-l` option makes `compare` lower-case the Ronin results before comparing them to the expected results.
 
 Use the `-h` help option on `compare` to find out more about the command-line options it understands.
-
-Tests for `py.test`
--------------------
-
-The files whose names begin with `test_` are intended for use with the [py.test](https://pytest.org) testing framework.  You can run them by executing the following command in a terminal shell in this directory:
-
-```csh
-python3 -m py.test
-```
