### 2018-03-26 #646 @mhucka
@arfon Similar to my question related to another paper recently submitted: I neglected to include an NSF agency funding acknowledgment in this paper as well. Can I go ahead and modify the paper that's in the github repo, or is there a different procedure I should follow?
