### 2018-03-27 #646 @arfon
@timClicks, @ahurriyetoglu - would you be willing to review this submission for JOSS? As this submission is outside of my research area I would like to find two reviewers for this submission.
