### 2018-03-27 #646 @timClicks
I've taken some time to review the paper and the project. Well done, it's very impressive. Thanks for taking the time to refine so many details.

My decision is to **Accept** the paper, subject to clarification about the first point below.

Here my review notes with reference to the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines):

- The "JOSS paper" section requires a list of authors and affiliations. This paper doesn't include an explicit list, which I assume is yourself as a single developer. @arfon - could you please provide guidance as to whether the paper's listed author is sufficient?
- license - ✔️ - spiral provides a copy of the [GPL v3 license](https://github.com/casics/spiral) within the root of the distribution  (Acceptable)
- documentation ✔️
   - statement of need  ✔️ - ("The need for splitting identifiers arises in a variety of contexts, including natural language processing (NLP) methods applied to source code analysis and program comprehension. ")
   - installation instrutions ✔️ - spiral is installable via `pip`, which is documented in the `README` file ("`sudo pip3 install git+https://github.com/casics/spiral.git`")
    - dependencies ✔️ - described in a machine-readable format by the [`requirements.txt`](https://github.com/casics/spiral/blob/master/requirements.txt) file (Good)
    - example usage: ✔️
    - API documentation: ✔️ All public APIs such as the `spiral.ronin`, `spiral.samurai` and `spiral.simple_splitters` module include extensive, well-written documentation provided as docstrings that includes an overview of the API and the parameters in detail (Good)
- tests: [spiral](https://github.com/casics/spiral/tree/master/tests) includes a thorough test suite and includes instructions on how to run it
- community guidelines: ✔️ are available in the `README` files, with sections for developers and end-users

Here is some other feedback from me personally:

- excellent in-source documentation: the [source files](https://github.com/casics/spiral/blob/master/spiral/ronin.py#L202) contain extensive notes about implementation decisions
- the repository includes extensive details of the differences/limitations between splitting implementations
- cites prior art - spiral includes [references to further material](https://github.com/casics/spiral#-more-information), including alternative implementations
