### 2018-03-27 #646 @mhucka
Thanks for your review! About this item:

> The "JOSS paper" section requires a list of authors and affiliations. This paper doesn't include an explicit list, which I assume is yourself as a single developer.

I am confused about what this refers to. The paper currently includes an explicit list of authors and affiliations in the format defined by JOSS. Also, a quick comparison to other JOSS papers shows that the amount of information about the sole author provided in this paper (about Spiral) is identical to the amount of information provided in other published JOSS papers. Can you please clarify what you mean or are looking at when you say the paper doesn't include an explicit list? I'll be happy to correct the problem if possible.
