### 2018-03-27 #653 @timClicks
@arfon I've reviewed the paper & code and could only find this technicality preventing an Accept:

- **References**: My understanding is that WordNet should be given a formal citation in the paper https://wordnet.princeton.edu/citing-wordnet. The DOI for the Miller article is [10.1145/219717.219748](http://dx.doi.org/10.1145/219717.219748)

Some more general notes:

- Has the Ronin algorithm been described in its own right? It looks like a very interesting contribution to the state of the art.
- The repo is _large_ as it contains data files and installation takes a long time to download the ~70MB of assets. `pip` doesn't report progress made by `git` during the clone stage and initially believed that the process had hung when I ran `pip install git+https://github.com/casics/spiral.git` myself.
