+++ spiral/constants.py
@@ -21,6 +21,7 @@ symbols such as "int32", "float64", etc.
 # - https://en.wikipedia.org/wiki/List_of_codecs
 # - https://en.wikipedia.org/wiki/Comparison_of_video_codecs
 # - the book 'The Definitivei Guide to iReport' by Toffoli, 2007
+# - NIST Interagency Report 7581 (Sep. 2009)
 # - various lists and dictionaries of computing terms found by Googling
 # - personal experience
  
@@ -31,6 +32,9 @@ common_terms_with_numbers = {
     '2nd',
     '2of7',
     '3d',
+    '3des',
+    '3gpp2',
+    '3gpp',
     '3of9',
     '3ivx',
     '3rd',
@@ -49,6 +53,7 @@ common_terms_with_numbers = {
     'b2e',
     'b8zs',
     'base64',
+    'bgp4',
     'bit7',
     'bzip2',
     'bz2',
@@ -80,6 +85,7 @@ common_terms_with_numbers = {
     'db2',
     'db9',
     'des3',
+    'dhcpv6',
     'ds0',
     'ds1',
     'ean128',
@@ -87,6 +93,9 @@ common_terms_with_numbers = {
     'e2e',
     'e3cp',
     'ec2',
+    'ec2n',
+    'ext2fs',
+    'ext3fs',
     'f2f',
     'f95',
     'fat32',
@@ -131,6 +140,7 @@ common_terms_with_numbers = {
     'i386',
     'ia32',
     'ia64',
+    'ic3',
     'id10t',
     'ie10',
     'ie11',
@@ -169,6 +179,7 @@ common_terms_with_numbers = {
     'jdk15',
     'jdk16',
     'jpeg2000',
+    'jtc1',
     'koi8r',
     'l10n',
     'l2f',
@@ -178,6 +189,7 @@ common_terms_with_numbers = {
     'l3f',
     'l3s',
     'l3tp',
+    'l3vpn',
     'log10',
     'log2',
     'log4j',
@@ -212,12 +224,15 @@ common_terms_with_numbers = {
     'p4m',
     'pep8',
     'perl5',
+    'pkmv1',
+    'pkmv2',
     'pop3',
     'px64',
     'python2',
     'python3',
     'qt4',
     'r2d2',
+    'rc4',
     'rc5',
     'rc6',
     'rfc[0-9]{3,}',
@@ -271,6 +286,8 @@ common_terms_with_numbers = {
     'windows7',
     'windows10',
     'windows11',
+    'wpa2',
+    'wss4j',
     'x11',
     'x11r4',
     'x11r5',
