### 2018-03-30 #653 @mhucka
@timClicks Thank you for your review and comments.

With respect to citing WordNet, although I interpret their instructions to only require citing where WordNet is discussed (note the paper doesn't mention it), I think I should modify the paper to mention WordNet and then cite the recommended sources. In any case, the website doesn't properly acknowledge the papers (though it referenced WordNet and its license). I'll update that too.

With respect to the other questions:
* Ronin has not been described elsewhere. I hope to have time to write a separate paper about the algorithm.
* Turns out the problem is not so much Spiral itself (which totals 23Mb) but rather the git history. I will try to prune it using BFG – maybe we reduce the size.
