### 2018-03-30 7438840
Add a link to WordNet that is more in line with WordNet's guidelines
See https://wordnet.princeton.edu/citing-wordnet


Former-commit-id: d225c763e689538f372c69193e694759fe4ab8cd
