### 2018-03-30 75d2b24
Add citations for WordNet & NLTK; minor editing too
I edited the first paragraph to try to save a line, to avoid a bad
paragraph break that was caused by the introduction of the NLTK and
WordNet citations on the bottom of page 1.


Former-commit-id: e0ce70da0ce634cb5d4ede39e2a0a48cb6b480b8
