### 2018-03-30 #653 @mhucka
The new version of the paper has the new citations.
