### 2018-03-30 #653 @mhucka
Regarding the size: I've now spent an hour trying various git pruning methods, and while they definitely did remove some old cruft in the git history, the cloned repo size has not changed – it's still 70M.  I'm very annoyed, but I'm also at the limit of what I know how to do with git in reducing sizes. My best guess here is that the repeated updates of the data files are in git's history, and that's what's blowing up the size of the repo.

How much of a problem is the size issue? At this point, I think the only option to avoid it is to create a pypi distribution and avoid the git clone approach, which is something I wanted to do eventually anyway, but preferably not right now due to time constraints ...
