### 2018-04-01 #653 @arfon
> How much of a problem is the size issue? At this point, I think the only option to avoid it is to create a pypi distribution and avoid the git clone approach, which is something I wanted to do eventually anyway, but preferably not right now due to time constraints ...

@mhucka - I think this is OK and shouldn't be a blocker for accepting this submission.

At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
