### 2018-04-04 18665a7
Add Zenodo DOI and update release badge
