### 2018-04-04 #653 @mhucka
Thanks! This is now done. The DOI at Zenodo is [10.5281/zenodo.1211835](https://zenodo.org/record/1211835)
